carslisting/requirements.txt
django >= 3.2, <= 4.1
djangorestframework >=3.12, <= 3.13
psycopg2-binary==2.8.6

